Move clases and rename them.
