+++ compiler/src/dotty/tools/dotc/transform/linker/CollectSummaries.scala
@@ -5,7 +5,6 @@ import dotty.tools.dotc.ast.Trees._
 import dotty.tools.dotc.ast.tpd
 import dotty.tools.dotc.core.Constants.Constant
 import dotty.tools.dotc.core.Contexts._
-import dotty.tools.dotc.core.Decorators._
 import dotty.tools.dotc.core.Flags._
 import dotty.tools.dotc.core.Names._
 import dotty.tools.dotc.core.StdNames.nme
@@ -14,7 +13,8 @@ import dotty.tools.dotc.core.TypeErasure
 import dotty.tools.dotc.core.Types._
 import dotty.tools.dotc.transform.SymUtils._
 import dotty.tools.dotc.transform.TreeTransforms._
-import dotty.tools.dotc.transform.linker.summaries.{CallInfo, MethodSummary, OuterTargs}
+import dotty.tools.dotc.transform.linker.callgraph.OuterTargs
+import dotty.tools.dotc.transform.linker.summaries.{CallInfo, MethodSummary}
 import dotty.tools.dotc.transform.linker.types.{ClosureType, PreciseType}
 import dotty.tools.dotc.typer.Applications._
 
@@ -582,71 +582,5 @@ class CollectSummaries extends MiniPhase { thisTransform =>
 object CollectSummaries {
 
   def isPhaseRequired(implicit ctx: Context): Boolean = BuildCallGraph.isPhaseRequired
-
-  class SubstituteByParentMap(substMap: OuterTargs)(implicit ctx: Context) extends DeepTypeMap()(ctx) {
-    def apply(tp: Type): Type = {
-      lazy val substitution = substMap.mp.getOrElse(tp.typeSymbol.owner, Nil)
-      def termTypeIfNeed(t: Type): Type = {
-        if (tp.isInstanceOf[TermType] && !t.isInstanceOf[TermType]) {
-          t match {
-            case t: TypeAlias =>
-              assert(t.underlying.isInstanceOf[TermType])
-              t.underlying
-            case t: ClassInfo =>
-              t.typeRef
-            case _ =>
-              assert(false)
-              null
-          }
-        } else t
-      }
-      tp match {
-        case tp: RefinedType => mapOver(tp) // otherwise we will loose refinement
-        case tp: TypeAlias => mapOver(tp) // map underlying
-        case _ if tp.typeSymbol.exists && substitution.nonEmpty =>
-          var typ = tp
-          /*val id = tp.typeSymbol.owner.info match {
-            case t: PolyType =>
-              t.paramNames.indexOf(tp.typeSymbol.name)
-            case t: ClassInfo =>
-              var typ = tp
-              var id = t.typeParams.indexOf(typ.typeSymbol)
-              while (id < 0 && (tp.typeSymbol.info.typeSymbol ne tp.typeSymbol)) {
-                typ = tp.typeSymbol.info
-                id = t.typeParams.indexOf(typ.typeSymbol)
-              }
-              id
-            case _ =>
-              -2
-          } */
-          var id = substitution.find(x => x._1 == tp.typeSymbol.name)
-          var limit = 30
-          var stack: List[Type] = Nil
-          while (id.isEmpty && (limit > 0) && (typ.typeSymbol.info.typeSymbol ne typ.typeSymbol)) {
-            typ = typ.typeSymbol.info
-            stack = typ :: stack
-            id = substitution.find(x => x._1 == typ.typeSymbol.name)
-            limit -= 1
-          }
-
-          // assert(id.isDefined)
-          if (id.isDefined) {
-            val t = termTypeIfNeed(id.get._2.stripTypeVar)
-            if (!(t =:= typ))
-             apply(termTypeIfNeed(t))
-            else t
-          } else tp
-        case t: TypeRef if t.prefix.normalizedPrefix eq NoPrefix =>
-          val tmp = apply(t.info)
-          if (tmp ne t.info) termTypeIfNeed(tmp)
-          else mapOver(t)
-        case _ => mapOver(tp)
-
-      }
-
-    }
-  }
-
-  def substName = "substituted".toTypeName
 
 }
